Crate hydradx_traits

Re-exports

pub use registry::*;

pub use oracle::*;

Modules

liquidity_mining

nft

oracle

pools

registry

router

Structs

AMMTransfer

Hold information to perform amm transfer Contains also exact amount which will be sold/bought

Traits

AMM

Traits for handling AMM Pool trades.

AMMPosition

Implementers of this trait provides information about user’s position in the AMM pool.

CanCreatePool

LockedBalance

OnCreatePoolHandler

Handler used by AMM pools to perform some tasks when a new pool is created.

OnLiquidityChangedHandler

Handler used by AMM pools to perform some tasks when liquidity changes outside of trades.

OnTradeHandler

Handler used by AMM pools to perform some tasks when a trade is executed.

Resolver