Trait hydradx_traits::registry::AccountIdFor

pub trait AccountIdFor<Assets> {
    type AccountId;
    fn from_assets(assets: &Assets, identifier: Option<&str>) -> Self::AccountId;

    fn name(assets: &Assets, identifier: Option<&str>) -> Vec<u8>;
}

Abstraction over account id and account name creation for Assets

Associated Types

type AccountId

Required methods

fn from_assets(assets: &Assets, identifier: Option<&str>) -> Self::AccountId

Create account id for given assets and an identifier

fn name(assets: &Assets, identifier: Option<&str>) -> Vec<u8>

Create a name to uniquely identify given assets.