Trait hydradx_traits::OnTradeHandler

pub trait OnTradeHandler<AssetId, Balance> {
    fn on_trade(
        asset_a: AssetId, 
        asset_b: AssetId, 
        amount_in: Balance, 
        amount_out: Balance, 
        liq_amount: Balance
    );

    fn on_trade_weight() -> Weight;
}

Handler used by AMM pools to perform some tasks when a trade is executed.

Required methods

fn on_trade(
asset_a: AssetId,
asset_b: AssetId,
amount_in: Balance,
amount_out: Balance,
liq_amount: Balance
)

Include a trade in the average price calculation of the price-oracle pallet.

fn on_trade_weight() -> Weight

Known overhead for a trade in on_initialize/on_finalize. Needs to be specified here if we don’t want to make AMM pools tightly coupled with the price oracle pallet, otherwise we can’t access the weight. Add this weight to an extrinsic from which you call on_trade.