Trait hydradx_traits::AMM

pub trait AMM<AccountId, AssetId, AssetPair, Amount: Zero> {

    fn exists(assets: AssetPair) -> bool;
    fn get_pair_id(assets: AssetPair) -> AccountId;
    fn get_share_token(assets: AssetPair) -> AssetId;
    fn get_pool_assets(pool_account_id: &AccountId) -> Option<Vec<AssetId>>;
    fn get_spot_price_unchecked(
        asset_a: AssetId,
        asset_b: AssetId,
        amount: Amount
    ) -> Amount;
    fn validate_sell(
        origin: &AccountId,
        assets: AssetPair,
        amount: Amount,
        min_bought: Amount,
        discount: bool
    ) -> Result<AMMTransfer<AccountId, AssetId, AssetPair, Amount>, DispatchError>;
    fn execute_sell(
        transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Amount>
    ) -> DispatchResult;
    fn validate_buy(
        origin: &AccountId,
        assets: AssetPair,
        amount: Amount,
        max_limit: Amount,
        discount: bool
    ) -> Result<AMMTransfer<AccountId, AssetId, AssetPair, Amount>, DispatchError>;
    fn execute_buy(
        transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Amount>
    ) -> DispatchResult;
    fn get_min_trading_limit() -> Amount;
    fn get_min_pool_liquidity() -> Amount;
    fn get_max_in_ratio() -> u128;
    fn get_max_out_ratio() -> u128;
    fn get_fee(pool_account_id: &AccountId) -> (u32, u32);

    fn sell(
        origin: &AccountId,
        assets: AssetPair,
        amount: Amount,
        min_bought: Amount,
        discount: bool
    ) -> DispatchResult { ... }
    fn buy(
        origin: &AccountId,
        assets: AssetPair,
        amount: Amount,
        max_limit: Amount,
        discount: bool
    ) -> DispatchResult { ... }

}

Traits for handling AMM Pool trades.

Required Methods

fn exists(assets: AssetPair) -> bool

Check if both assets exist in a pool.

fn get_pair_id(assets: AssetPair) -> AccountId

Return pair account.

fn get_share_token(assets: AssetPair) -> AssetId

Return share token for assets.

fn get_pool_assets(pool_account_id: &AccountId) -> Option<Vec<AssetId>>

Return list of active assets in a given pool.

fn get_spot_price_unchecked(
asset_a: AssetId,
asset_b: AssetId,
amount: Amount
) -> Amount

Calculate spot price for asset a and b.

fn validate_sell(
origin: &AccountId,
assets: AssetPair,
amount: Amount,
min_bought: Amount,
discount: bool
) -> Result<AMMTransfer<AccountId, AssetId, AssetPair, Amount>, DispatchError>

Sell trade validation Perform all necessary checks to validate an intended sale.

fn execute_sell(
transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Amount>
) -> DispatchResult

Execute buy for given validated transfer.

fn validate_buy(
origin: &AccountId,
assets: AssetPair,
amount: Amount,
max_limit: Amount,
discount: bool
) -> Result<AMMTransfer<AccountId, AssetId, AssetPair, Amount>, DispatchError>

Buy trade validation Perform all necessary checks to validate an intended buy.

fn execute_buy(
transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Amount>
) -> DispatchResult

Execute buy for given validated transfer.

fn get_min_trading_limit() -> Amount

fn get_min_pool_liquidity() -> Amount

fn get_max_in_ratio() -> u128

fn get_max_out_ratio() -> u128

fn get_fee(pool_account_id: &AccountId) -> (u32, u32)

Provided Methods

fn sell(
origin: &AccountId,
assets: AssetPair,
amount: Amount,
min_bought: Amount,
discount: bool
) -> DispatchResult

Perform asset swap. Call execute following the validation.

fn buy(
origin: &AccountId,
assets: AssetPair,
amount: Amount,
max_limit: Amount,
discount: bool
) -> DispatchResult

Perform asset swap.